Move shared code to `rest/data/`
