+++ graphql/util.go
@@ -1253,47 +1253,3 @@ var projectPermissionAccessMap = map[ProjectPermission]map[AccessLevel]Permissio
 		AccessLevelView: {Permission: evergreen.PermissionLogs, Level: evergreen.LogsView.Value},
 	},
 }
-
-const (
-	projectIdKey  = "projectId"
-	identifierKey = "identifier"
-
-	// Keys that are used in the paramsMap.
-	projectIdentifierKey = "projectIdentifier"
-	repoIdKey            = "repoId"
-	versionIdKey         = "versionId"
-	patchIdKey           = "patchId"
-	taskIdKey            = "taskId"
-)
-
-func buildProjectParameterMap(args map[string]interface{}) (map[string]string, error) {
-	paramsMap := map[string]string{}
-
-	if projectIdentifier, hasProjectIdentifier := args[projectIdentifierKey].(string); hasProjectIdentifier {
-		paramsMap[projectIdentifierKey] = projectIdentifier
-	}
-	if identifier, hasIdentifier := args[identifierKey].(string); hasIdentifier {
-		paramsMap[projectIdentifierKey] = identifier
-	}
-	if projectId, hasProjectId := args[projectIdKey].(string); hasProjectId {
-		paramsMap[projectIdentifierKey] = projectId
-	}
-	if repoId, hasRepoId := args[repoIdKey].(string); hasRepoId {
-		paramsMap[repoIdKey] = repoId
-	}
-	if versionId, hasVersionId := args[versionIdKey].(string); hasVersionId {
-		paramsMap[versionIdKey] = versionId
-	}
-	if patchId, hasPatchId := args[patchIdKey].(string); hasPatchId {
-		paramsMap[patchIdKey] = patchId
-	}
-	if taskId, hasTaskId := args[taskIdKey].(string); hasTaskId {
-		paramsMap[taskIdKey] = taskId
-	}
-
-	if len(paramsMap) == 0 {
-		return nil, errors.New("params map is empty")
-	}
-
-	return paramsMap, nil
-}
