+++ rest/route/middleware.go
@@ -10,13 +10,10 @@ import (
 
 	"github.com/evergreen-ci/evergreen"
 	"github.com/evergreen-ci/evergreen/model"
-	"github.com/evergreen-ci/evergreen/model/build"
 	"github.com/evergreen-ci/evergreen/model/distro"
 	"github.com/evergreen-ci/evergreen/model/event"
 	"github.com/evergreen-ci/evergreen/model/host"
-	"github.com/evergreen-ci/evergreen/model/patch"
 	"github.com/evergreen-ci/evergreen/model/task"
-	"github.com/evergreen-ci/evergreen/model/testlog"
 	"github.com/evergreen-ci/evergreen/model/user"
 	"github.com/evergreen-ci/evergreen/rest/data"
 	"github.com/evergreen-ci/evergreen/util"
@@ -614,98 +611,6 @@ func RequiresSuperUserPermission(permission string, level evergreen.PermissionLe
 	return gimlet.RequiresPermission(opts)
 }
 
-func GetProjectIdForProjectScopes(ctx context.Context, paramsMap map[string]string) (string, int, error) {
-	projectID := util.CoalesceString(paramsMap["projectId"], paramsMap["projectIdentifier"])
-	repoID := paramsMap["repoId"]
-
-	var err error
-	versionID := paramsMap["versionId"]
-	if projectID == "" && versionID != "" {
-		projectID, err = model.FindProjectForVersion(versionID)
-		if err != nil {
-			return "", http.StatusNotFound, errors.Wrapf(err, "finding version '%s'", versionID)
-		}
-	}
-
-	patchID := paramsMap["patchId"]
-	if projectID == "" && patchID != "" {
-		if !patch.IsValidId(patchID) {
-			return "", http.StatusBadRequest, errors.New("not a valid patch ID")
-		}
-		projectID, err = patch.FindProjectForPatch(patch.NewId(patchID))
-		if err != nil {
-			return "", http.StatusNotFound, errors.Wrapf(err, "finding project for patch '%s'", patchID)
-		}
-	}
-
-	buildID := paramsMap["buildId"]
-	if projectID == "" && buildID != "" {
-		projectID, err = build.FindProjectForBuild(buildID)
-		if err != nil {
-			return "", http.StatusNotFound, errors.Wrapf(err, "finding project for build '%s'", buildID)
-		}
-	}
-
-	testLog := paramsMap["logId"]
-	if projectID == "" && testLog != "" {
-		var test *testlog.TestLog
-		test, err = testlog.FindOneTestLogById(testLog)
-		if err != nil {
-			return "", http.StatusInternalServerError, errors.Wrapf(err, "finding test log '%s'", testLog)
-		}
-		if test == nil {
-			return "", http.StatusNotFound, errors.Errorf("test log '%s' not found", testLog)
-		}
-		projectID, err = task.FindProjectForTask(test.Task)
-		if err != nil {
-			return "", http.StatusNotFound, errors.Wrapf(err, "finding project for task '%s' associated with test log '%s'", test.Task, test.Id)
-		}
-	}
-
-	taskID := paramsMap["taskId"]
-	if projectID == "" && taskID != "" {
-		projectID, err = task.FindProjectForTask(taskID)
-		if err != nil {
-			return "", http.StatusNotFound, errors.Wrapf(err, "finding project for task '%s'", taskID)
-		}
-	}
-
-	if repoID != "" {
-		var repoRef *model.RepoRef
-		repoRef, err = model.FindOneRepoRef(repoID)
-		if err != nil {
-			return "", http.StatusInternalServerError, errors.Wrap(err, "finding repo")
-		}
-		if repoRef == nil {
-			return "", http.StatusNotFound, errors.Errorf("repo '%s' not found", repoID)
-		}
-		return repoID, http.StatusOK, nil
-	}
-
-	// Return an error if the project isn't found
-	if projectID == "" {
-		return "", http.StatusNotFound, errors.New("no project found")
-	}
-
-	projectRef, err := model.FindMergedProjectRef(projectID, versionID, true)
-	if err != nil {
-		return "", http.StatusInternalServerError, errors.Wrap(err, "finding project")
-	}
-	if projectRef == nil {
-		return "", http.StatusNotFound, errors.Errorf("project '%s' not found", projectID)
-	}
-	usr := gimlet.GetUser(ctx)
-	if usr == nil && projectRef.IsPrivate() {
-		return "", http.StatusUnauthorized, errors.New("unauthorized")
-	}
-
-	if projectRef.Id == "" {
-		return "", http.StatusInternalServerError, errors.New("project ID is blank")
-	}
-
-	return projectRef.Id, http.StatusOK, nil
-}
-
 func urlVarsToProjectScopes(r *http.Request) ([]string, int, error) {
 	var err error
 	vars := gimlet.GetVars(r)
@@ -732,7 +637,7 @@ func urlVarsToProjectScopes(r *http.Request) ([]string, int, error) {
 		"taskId":    util.CoalesceStrings(append(query["task_id"], query["taskId"]...), vars["task_id"], vars["taskId"]),
 	}
 
-	projectId, statusCode, err := GetProjectIdForProjectScopes(r.Context(), paramsMap)
+	projectId, statusCode, err := data.GetProjectIdFromParams(r.Context(), paramsMap)
 	if err != nil {
 		return nil, statusCode, err
 	}
